requirements.txt
theseus-ai 
tqdm 
configargparse 
h5py 
pyglet 
scikit-learn 
trimesh 
matplotlib
tensorboard
natsort 
pytorch3d
addict>=2.4.0
certifi
diffusers[torch]
einops
ninja
opencv-python-headless
pandas>=1.5.1
Pillow>=9.2.0
pyrender
scipy>=1.9.0
seaborn>=0.12.1
shapely>=2.0.0
six>=1.16.0
torch==2.0.1 -e https://download.pytorch.org/whl/cu118
torchvision==0.15.2 -e https://download.pytorch.org/whl/cu118
torchaudio==2.0.2 -e https://download.pytorch.org/whl/cu118
# wandb==0.13.6 # Logging
yapf==0.32.0
pytorch-lightning